changed buad_divint bits value to u32

+++ src/uart0/uartibrd.rs
@@ -56,8 +56,8 @@ pub struct BAUD_DIVINT_W<'a> {
 impl<'a> BAUD_DIVINT_W<'a> {
     #[doc = r"Writes raw bits to the field"]
     #[inline(always)]
-    pub unsafe fn bits(self, value: u16) -> &'a mut W {
-        self.w.bits = (self.w.bits & !0xffff) | (value as u32 & 0xffff);
+    pub unsafe fn bits(self, value: u32) -> &'a mut W {
+        self.w.bits = (self.w.bits & !0xffff) | (value & 0xffff);
         self.w
     }
 }
